fix: use decimal quantities for raw material stock deduction
Root cause: recipe ingredient quantities (e.g. 0.25 kg) were being
rounded up via int(ceil(...)) before deducting stock, causing 0.25 kg
to always deduct as 1 unit.

Changes:
- stock.quantity and stock_ledger.quantity changed to NUMERIC(12,3)
- adjust_stock() now accepts float instead of int
- _build_recipe_requirements() returns dict[int, Decimal] (no rounding)
- inventory /add and /remove routes accept float qty
- table_billing: remove int() casts around recipe requirements
- Mobile InventoryScreen: fix subtract endpoint to /remove (was 404)
- Migration: 20260417_decimal_stock_quantity.sql

+++ backend/app/models/stock.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from app.db import Base
 
@@ -12,7 +12,7 @@ class Inventory(Base):
 
     branch_id = Column(Integer, ForeignKey("branch.branch_id"))
 
-    quantity = Column(Integer, default=0)
+    quantity = Column(Numeric(12, 3), default=0)
     min_stock = Column(Integer, default=0)
 
     item = relationship("Item")
